theWolves/requirements.txt
certifi==2018.8.24
Django==2.1django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-filter==2.0.0
djangorestframework==3.8.2
django-tables2==2.0.1
Markdown==2.6.11
mysqlclient==1.3.12
mysql-connector-python==8.0.12
pip==18.0
pytz==2018.5
setuptools==40.2.0
wheel==0.31.1
